Revert ILRepack introduced by #24 (#46)
The ILRepack has introduced type conflict issues that are not easily
solved. This boils down to the fact that the public API surface of the
Platform SDK exposes types provided by dependencies. ILRepack creates an
artificial assembly to export those types.

However any user will need to import the original dependency assembly in
order to construct function arguments with the types in question. This
will result in duplicate type errors as they will exist in both the
user-imported dependency and the new artificial assembly created by the
ILRepack.